On-the-spot preferences
Allow for handling of resetting multiple preferences individually.

Present the user with a list of active on-the-spot preferences to
choose from.
If a preference dialog was never presented to the user, it will not
be available for reset.
If only one preference is currently active, a simple reset dialog is
presented (current existing dialog / no list)

Change-Id: If9c483e0745000ab486f060bf87756204b4a8e6e