Themes: Port to CM13 [3/3]
Change-Id: Ia8e852e5e5ad2465d0b9b73f16c7080a7337bb8a

Themes: Allow switchbar text color to be themeable

Better exposing the color of the switch bar text so that a theme
can handle this assignment separately without having to touch
framework theme colors.

Change-Id: I81e3430e188f1ee0a22e92aea50878980f9cc61f

Break out the add icon to be separately themeable.

If Settings is using the accent color for app bar, the add icon
is also colored with this value typically and will cause readability
issues. Adding a new drawable name that can be referenced to allow
separate theming.

Change-Id: I20512091e7a36c510399cbc5caa73aa8817fe393