Allow hiding of the battery status bar indicator
This commit simply adds a special value for a hidden status bar
indicator. This works due to the fact that the existing indicators
check that the current value matches their one which means that if
no indicators will be visible if no matches are found. This is usable
by people who like minimalistic aproaches and also assists people who
modify their systems to add some special battery indicators that are
no provided by the default system (as otherwise they would end up with
two indicators side-by-side).

It should still be noted that without changes in the system UI the QS
tile for battery looks a bit weird when this is applied as the battery
indicator disappears from there as well.

Change-Id: Ie23382817e00a4a9a13115cba2db789d038910a7