Allow manual media scanning requests
This commit adds media rescanning to the developer options screen.
It is a bit of a hack, abusing the behavior that takes place when
the volume is mounted, but it works for our needs unless we want to
set up a connection with the media scanner service from the settings
application. We could use a progress indicator in the notification
drawer probably as this is a long running process, but that can be
left to be implemented another day.

Change-Id: Ifbc86465990d76ab2700c997d28e029559805d15