Settings: Implement Immersive mode
Simple menus are our new friends. With this commit, they just provide
what is essentially a stock/immersive toggle for the navigation bar
and the status bar, but can be easily expanded to support awesome
features such as Pie or [redacted] when the appropriate changes get
pushed to the frameworks.

Change-Id I1f361d59d2b621d62d21442830549591b687bab7 (in the
frameworks/base repository) should be merged prior to this
change as it provides the changes in the core system required
to make the simple menus buildable and functional.

Change-Id: I11859fd4197d636bbfff25844a56794b0a79372f