AppOps: Rename the App Ops category
Users might not all be familiar with the methodology of Application
Operations: the point of the feature is to allow the user to manage
what operations each application is allowed to carry out, that is,
which permissions are allowed and which are blocked. Due to that, we
can simply call this section as "Permissions" - as that is what the
end user will be managing, regardless of the backend implementation.

Change-Id: I754447c03417443a99092eb88d3ef9c8c3d79e49