MSIM: Fix Settings crash on sim enable/disable
Remove static for AlertDialog and ProgressDialog.
When Home key is pressed while sim enable/disable is
in-progress, SimSettings fragment will be destroyed & both
Dialogs will be detached from the fragment.
When fragment is recreated, IllegalStateException occurs
if older Dialog instances are accessed.
Display dialouges only when fragment is visible.

CRs-Fixed: 971548
Change-Id: I2059489decad9496a72903a60bf89b17b60fd4a8