FeaturePreferences: Introduce settings reset
This introduces a new selectable settings reset flow under
"Backup and reset." The user is presented with a list of
settings that were set via OTS and given the choice to reset
a select few of them or reset all of them. This will cause OTS
to show the setting choice snackbar again next time the feature
is used.

Change-Id: Ic7e787834d3bb01b2a6ce0a01ecaaf1a85b64d34
Signed-off-by: Evan Anderson <evananderson@aospa.co>