Update patch set 2
Patch Set 2: Code-Review+1

This is a better way of portraying this as it is accurate & on point. We shouldn't leave it in its current state as it is not precise. I do not understand the organization of other languages besides English (and some Spanish) so if others could translate the strings, then that would be awesome... Though I am only concerned with the English

Patch-set: 2
Label: Code-Review=+1