Update patch set 4
Patch Set 4:

That would be great, but I didn't really change much with this commit. I will do a runthrough of it all later. I just noticed this today and thought the message was a bit redundant along with the "plug" message so I made it useful

Patch-set: 4
Label: Code-Review=0