lights: Make light dialogs dismissible via notification
The Battery and Notification Light dialogs used to show
an ongoing notification indicating that the LED has been
enabled by settings. Since the Preference and Dialog
classes don't support any onPause() or onResume()
functions in their lifecycle it is not possible to
determine (outside the SettingsFragment) with certainty
whether the dialog is still on the screen.

When a user opens the dialog and presses the home button
the fragment is paused and the Dialog's onStop() method
is not called which causes ongoing notification to show
until manual intervention, namely until the dialog is
closed or until a new one has been opened.

To improve UX regarding this design issue we make the
notification dismissable and add a cancel button to it.

Change-Id: I6bacea3e0ed623b8eba8ed0a260f5db93db91f21
Signed-off-by: Chris Lahaye <dev@chrislahaye.com>