Update patch set 23
Patch Set 23: Code-Review-1

I really don't get it. In the previous PS it the switch was disabled and it was working, and in this PS it does the same only the preference doesn't stick. I didn't think flipping a switch to be off by default could get so complicated haha

Patch-set: 23
Label: Code-Review=-1