BillingCycleSettings: Fix two crash cases for the data usage warning
After entering a large number into the dialogue, some
digits will be converted to "E". "E" is not a valid
number, so it causes a crash.

Also a check for the decimal dot was added in
c05a3a02278c ("Fix Settings crash when setting
only decimal point to data usage warning")
However that check isn't sufficient, as the
decimal dot can be combined with a digit.

To fix both crashes, check whether the target string
contains the decimal dot or "E" and reset the value
to 0 for that case.

Test: none.
Change-Id: I6462bf510562eb3fb06304e50b68fba05d37b286
Signed-off-by: Alex Naidis <alex.naidis@linux.com>