Settings: add Device Name on About Phone
This will allow us to display the real commercial device name
on Settings so users doesn´t get confused. As well, this will
allow us to still tag our devices with their real device.model
so there won´t be any kind of app compatibilities at all.

As of that, we will be hiding device model in About Phone.

Change-Id: If105a616471bc33adcbfc0f20cda9823d4053681
Signed-off-by: Hernán Castañón <herna@paranoidandroid.co>