Settings: NPE occurs when SD card format as internal storage
Prompt NPE when format SD card as internal storage. SD remount at
the time formating, but app start a thread to read the volume info.
if the remount is not finish, the voluem info should be null.

So block the thread if get null, wait for the mounted operation finished.

Change-Id: Id5a677ea7367ad4f86d48f981f051371a13756f9
CRs-Fixed: 2171308