Change the layout of TP pages and previews
Change the position of the "Keep my wallpaper" checkbox and
make the bottom panel larger, as well as restrict how much
a preview card can grow or shrink depending on the screen
size.

Note that the UX spec cannot be met 100% on all screen
sizes, so for some cases we round to the closest
approximation.

(ConstraintLayout FTW!)

Fixes: 130763357
Fixes: 130398264
Change-Id: I15bece84fb4ce1018676f57d32e0fc533bcdec92