Update patch set 1
Patch Set 1:

should we be able to adjust the intensity of other vibrations, such as touch vibrations and alarms?  I can see where this would get complicated quickly for UX but I do know users asked for more advanced vibration settings so want to be sure we deliver everything that's expected.

Patch-set: 1
CC: Gerrit User 1000012 <1000012@3162e3d8-d894-470e-abd9-724f7b0effad>