Wi-Fi: Add 6GHz band support in Soft-AP UI advanced options
This changes adds support for following:

6GHz SAP: Add 6GHz band to the drop-down menu of Soft-AP band selection
switch to allow users to start Soft-Ap in 6GHz band. This option will not
be displayed if the device is not supporting 6GHz band.

This changes also removes wifi_ap_band_vendor_config_full and
wifi_ap_band_vendor_summary_full as now we are dynamically adding
SAP bands.

Change-Id: Id39a97d11624f5d8918aca4c1f7e8d88f5ef9f58
CRs-Fixed: 2569221