Add sanity check for slotIndex during scan abort
There is no sanity check for the slotIndex passed on as a parameter to
abort an ongoing PLMN scan. This leads to unexpected results in
scenarios where the slotIndex changes, for example, during SIM removal.
Fix this by honouring an abort request only if the slotIndex is valid.

Also, add logs for activity events for the PLMN scan results activity.

Change-Id: Ib515c04a22b521bab8a27e3eb5f90a31235fe262
CRs-Fixed: 2673149