Update patch set 1
Patch Set 1:

If we're going to move it to the searchbar it seems kind of redundant, but it can stay if you want. The avatar in About Phone doesn't really do anything. It doesn't show your actual avatar and the blue default avatar can't be themed at all, and even when it does it's often the wrong color opposed to what you have set

Patch-set: 1
Reviewer: Gerrit User 1000447 <1000447@3162e3d8-d894-470e-abd9-724f7b0effad>