Disable SIM On/off option during call
Currently SIM On/Off option is disabled only for SIM on which
call is active, user is allowed to toggle SIM option for the other
SIM. SIM On/Off request triggered during call is delayed at modem
due to active call. This leads to ANR in Settings app and unwanted
back to back requests

Fix is to block SIM On/Off option on both the SIM's when call is
started. Replace phonestatelistener with call state update intent
and read call state from TelephonyManager whenever intent is received

Change-Id: I91f8f76bd4107c3e6b3d619534ab947a47440d8b
CRs-Fixed: 2809313