[WIP] Settings: Add preference for KEY_MIN_REFRESH_RATE
Change-Id: Iac1f65ab09717ea55a5b471e094385c77ba894ee
Signed-off-by: chrisw444 <wandersonrodriguesf1@gmail.com>
Signed-off-by: Enyer <mamutal91@gmail.com>
Signed-off-by: chrisw444 <wandersonrodriguesf1@gmail.com>

[chrisl7]:

We are currently facing a CLO issue where Dinamyc Smooth Display is causing high flicker on AMOLED/OLED displays;
Then Add this temporary feature so that the user can define the minimum rate that should be used. So the user will not be forced to stay with 90hz or 120hz all the time.

Change-Id: Ie988e09e053e4d41791ccad15577c47731e9ff88
Signed-off-by: chrisl7 <wandersonrodriguesf1@gmail.com>