Disable DDS option and SIM switch during ECBM/SCBM
Like ECBM, when the device is under SCBM, requests for DDS switch and
SIM on/off are rejected. Grey out the related preferences on the UI when
such state is detected.

Also, fix error in reading ECBM state by switching to SDK APIs instead
of the restricted system properties.

Change-Id: Ice1b636cb7d30b4a909c5f8b3462c95e75a813ad
CRs-Fixed: 3097283