Dismiss user confirmation dialog only for physical sim
Currently a user confirmation dialog will be displayed both
for enable and disable eSIM. But for physical sim, dialog is
generally shown only while disabling physical sim.

If physical sim is removed while user is in the disable confirmation
dialog, the dialog is dismissed. This dismissal of dialog is not
applicable for eSIM, because eSIM is non-removable.

There is a conflict now when user tries to enable eSIM, user
confirmation dialog is dismissed resulting in failure to enable eSIM.

Fix: Check for whether the sim is embedded, before dismissing user
confirmation dialog.

Change-Id: I558a014932e90ba5e94a47215d4f2475125435e7
CRs-Fixed: 3248592