SAP: fix Dual Band display issue in Enhanced Open security
In UI there is a single option to allow Enhanced Open
soft AP configuration. However, internally, for band
2.4GHz/5GHz, OWE transition mode is used and for 6GHz
OWE only mode is used.

Currently, when Enhanced Open option is selected in UI,
band option display will be updated only when security
is OWE transition and band index is BAND_BOTH_2G_5G. As
a result, Dual Band option is not allowed but it will
still be displayed. Remove check for BAND_BOTH_2G_5G to
fix this and also add check for OWE for same behavior.

CRs-Fixed: 3286756
Change-Id: Ifae4b223009fa09682599fca08743ed62d80ae19