Show C_IWLAN warning for DSSS mobile data disable case
If device is in a C_IWLAN call, and the user tries to disable mobile
data, a warning dialog is shown to the user by design. This dialog is
applicable only when the device has at most one active subscription.

The existing code does not consider the case when the SIM configuration
is set to multisim but with only one SIM card inserted. Make changes so
that the dialog gets shown for such cases as well.

Change-Id: I2eabf1c51a01ab83a6357d9e7f9c07d60bbec4c5
CRs-Fixed: 3352941