Fix warning dialog not showing for data toggle
- In C_IWLAN-only mode, a warning dialog must be shown to the user when
they try to turn off the mobile data toggle during a call. Previously,
we relied on the DdsDataOptionStateTuner class to check if the DDS had
an ongoing voice call. This worked in the case of MSIM with only one
SIM enabled/inserted or SSIM. However, when both subs are active in the
case of MSIM, we will not be notified of the call state changes on the
DDS simply because we never registered to be notified of it.
- Prevent the warning dialog from being shown when changing the
preferred nw mode of the nDDS when it and the DDS are both in
C_IWLAN-only mode and the DDS is in a call.

Change-Id: Ib41974ab657a1e9f08cadb3f4963d3690fc0556c
CRs-Fixed: 3444725