Call A2dp, HFP disable profile for group.
For CSIP group devices, if first connected left or right earbud is
LE Audio only it will be considered as main device. If LE audio
profile is disabled, A2dp is enabled and active device will not
be the same main device. When media audio is disabled, profile
disable is called for main device which doesn't support a2dp
sink uuid. Media audio is grayed out but profile is not
disconnected. Streaming will still go over A2dp.

Fix is to call disable profile for the group, so disable will
be called for both left and right earbud. Earbud that supports
A2dp sink and profile connected will be disabled.

CRs-fixed: 3437488
Change-Id: I36615b33d34263d0154151d5fb345cf9e74da65e