Fix to enable eSIM for MEP supported UICCs
Currently the telephony framework identifies the
card as eSIM only when LPA app which implements
EuiccService is present.When QC LPA app is present
eSIM is getting detected as physical due to which
the eSIM enable operation at Settings is failing.
To fix it, check if sub is eSIM from the UiccSlot.

Change-Id: I90f06ad8736ab39e73c60abb2b168b9d5010eb26
CRs-Fixed: 3547917