TetherSettings: Use SoftApCallbacks for band settings.
Register SoftApCallback to listen for supported capabilities and update
the bands option in TetherSettings accordingly. List of bands and current
band on which softap is running is updated based on the callback.

This change is optimizing the WifiManager's getAllowedChannels Api calls
by 4 times.

Issue 1: UI was misbehaving as soon as SAP is turned on from Settings.
Issue 2: Mismatch in the UI with the Band preferences.

Earlier, Band frequencies were updating only once, there was a country
code change happened which resulted to the above issue. Now, band
preferences will be updated based on onCapabilityChanged() callback.

CRs-Fixed: 3701872
Change-Id: I33d7764cf30bd6ffab49e5a16e6a0e6f55da1552