Refactor MEID display function for About phone
To simplify the logic of the MEID display for some operator specific
requirement, add a preference sperately for CDMA so that the logic
gets clearer to understand the display order through the following
sequence:
    Slot1: MEID display
    Slot1: IMEI display
    Slot2: IMEI display
Besides, when CDMA isn't supported, MEID never needs to be displayed.
With this change, make sure above get achieved.

Change-Id: I18c55358a62ee73c580e8085b362a963ac98619f
CRs-Fixed: 4172466