SimRepository: use helpers to check for telephony support
Rely on the Utils.isMobileDataCapable() and Utils.isVoiceCapable()
methods of checking for telephony data and voice support instead of
looking for the FEATURE_TELEPHONY flag. This brings consistency to the
telephony support checks in the codebase.

The unit tests are updated to cover the different ways that Telephony
support can be disabled in the Settings application.

This does not change behavior of the app, it can be considered a
simple refactor, just to keep the codebase consistent.

Bug: 395714454
Test: Build and install in cellular-capable and cellular-less devices.
Test: atest SimRepositoryTest
Test: atest MobileNetworkSettingsSearchIndexTest
Flag: EXEMPT PURE_REFACTOR

Change-Id: I88138d2dbde63cfbbff85b3abf5c181e931ae94f