Fixing TTS output for button order settings
Configures the radio button items such that only the top level items are focusable, not the labels or icons. The TTS output for the items should read out the contents of the labels, but not the contents of the icons, as when combined the icons contain the same text as the labels.

See the bug for a demo video.

Test: manually confirm above description using Talkback
Bug: 410599460
Flag: android.view.accessibility.navbar_flip_order_option
Change-Id: Ib1343a7c84ab02607c2f28486d1ecb30c451e993