Fix a crash in BubblePreference
If the view had been set up previously, the checkedChangeListener
could be set so when setChecked was called on the various prefs
it could result in an illegal state exception from recycler view.

The fix is to null out the listener, set the checked state, then set
the listener.

Flag: EXEMPT BUGFIX
Test: manual - change settings in bubble app settings page and make
               sure everything works
Bug: 443905341
Change-Id: I4a2b1564c7993a57b3bf8e2866b1b6cef5c919bb