Enlarge DisplayBlock bounds to create space for a11y arrow buttons
Bounds are enlarged while maintaining the existing bounds for Wallpaper,
Background and highlight. Please take a look at the screenshots below:
- Existing: http://shortn/_prhlatKEoV
- Enlarged bounds: http://shortn/_4MkUbYTlsT

Listener is also updated so that it keeps the trigger bounds to only be
on the display wallpaper

In summary, this doesn't have any visual or behavior change to the user.
The actual arrow buttons will be added in the next CL

Bug: 438649236
Test: atest DisplayBlockTest
Flag: EXEMPT BUGFIX
Change-Id: Ie6d1ef1cb75d8f5fd3b501b89e238fcb4fd76da9