[Dual Shade] Use a config resource value to show/hide the Shade Setting.
This uses SettingsLib's `config_useDualShadeSetting`, a shared config
value with SystemUI, to determine the visibility of the setting to
enable/disable Single/Dual shade modes.

This isn't merely a refactoring, but a change in behavior as well:
instead of relying on the screens available on the device, this relies
on the size of the screen in which the Settings app is shown. Since the
setting only applies to compact screens, it is only shows in the
Settings app on those screens.

Fix: 438813342
Bug: 444424242
Bug: 442479130
Test: Updated unit tests.
Test: Verified manually by opening the Settings app on both a folded and
 unfolded device, and observing that the Dual Shade setting is shown and
 hidden in both the Settings app and QS shade shortcut, respectively.
Flag: com.android.systemui.scene_container
Change-Id: I0de96d0eebf7b44f72b37e6a74147c80c9dd44a6