Calculate which direction to display arrows
This is implemented based on DisplayTopologyGraph edges.
- 1. Check if 2 displays are connected on a corner. If so, display
should be able to move on those directions
- 2. Check if any display is blocking on any direction. If there is,
mark that direction as blocked. At the same time, allow movement
perpendicular to the blocked display direction (e.g. if there's a
display is on the left / right, then the current display can move down
or up
- 3. Check if display can move to a particular direction, and not
blocked in step2 by any other displays, then this direction should be a
move option.

Demo: http://shortn/_yFjN0BB7jF

Note: This CL is not showing any arrows yet, the show/hide toggle is
implemented in the next CL

Bug: 438649236
Test: DisplayTopologyPreferenceA11yUtilsTest
Flag: EXEMPT BUGFIX
Change-Id: I217e32c56901874d453bb197e74430fe132a4686