Fix Settings search for Airplane Mode
When airplane mode sync is avaiable, the PreferenceController should not
be available. This is causing an extra button displayed in the Search
results. Changed to override indexable instead of the isIndexable() in
PreferenceIndexableProvider. Also since we're using different preference
keys to distinguish the two, to fix the highlighting, we provide the
right key in the launchIntent.

Bug: 450019907
Flag: com.android.server.connectivity.sync_airplane_mode_with_watches
Test: atest SettingsRoboTests
Change-Id: Ie5426450158dc7ee4fcfe87d4bf1b4558769b7f4