Update icons for share/allow edit toggles in wifi network details
Previously in the settings app, network node and edit icons were
used as icons for the shared network and allow edit toggles,
which weren't intuitive. Therefore, replace the icons with the
group icon and the network manage icons to make their function
clearer to the user. Refer to b/448707226 for the new mocks.

Bug: b/448707226
Flag: com.android.settings.connectivity.wifi_multiuser
Test: manual testing
Change-Id: I1aa867359bcb6c551caed85d9b5fd00c089547b4