Add test for homepage container focusability
Adds a new unit test to verify that the main homepage container view is not focusable. This test confirms the behavior resulting from the removal of code that previously set the container to be focusable and requested focus for it.



This CL is generated by AI, please check the comments for more details.

Test: ATP tests passed http://go/forrest-run/L19900030025614723
Bug: 431235865
Flag: TEST_ONLY



Change-Id: Ia437d5dc55570fc25fb0ea2d856a993fc0d04fad