Update rules when no carrier config in enabled network modes preference
If for any reason there is no carrier config available, it is assumed
that no restriction to the corresponding network mode exists.

E.g. if the updateConfig() method is called and there is no carrier
config, there is no way to get the KEY_PREFER_3G_VISIBILITY_BOOL
preference, so it will be assumed to be true (i.e. there is no carrier
restriction applied to hide the 3G preference).

This change also reorders the rules used to display the different
network modes, starting from 2G and up to 5G, also printing in the log
the necessary information to understand why a given item may not have
been displayed.

NO_IFTTT=No corresponding change required

Bug: 449919991
Bug: 449910704
Test: atest EnabledNetworkModePreferenceControllerTest
Flag: EXEMPT BUGFIX

Change-Id: Ibacc840b92b4ebd69603bc59f16d59a422ae5f4d