Ensure unsupported network modes are not displayed
We can query from TelephonyManager the list of supported network
modes, and then use that information to decide whether the item should
be shown or not. This was already done for 5G, now it is done for all
2G, 3G and 4G as well.

The bitmask for 3G includes the deprecated 3GPP2 modes (e.g. EVDO)
only for completeness. This preference controller should eventually
remove the 3GPP2 support altogether.

NO_IFTTT=No corresponding change required

Bug: 449919991
Bug: 449910704
Test: atest EnabledNetworkModePreferenceControllerTest
Flag: EXEMPT BUGFIX

Change-Id: I949d39f02bf8735a923b036c088a2f63bed5b450