Hide 2G network protection in devices without 2G support
There is no point in allowing the user to configure the 2G network
protection if the modem itself doesn't support 2G, so make sure the
section (Enable2gPreferenceController) is not displayed on devices
that don't support 2G.

The "2G network protection" section is used inside each SIM settings,
and also in the "Mobile Network Security" section (under "Security &
Privacy"). In the latter (CellularSecurityPreferenceController),
checking for 2G support must also be done to decide whether the
whole section is displayed or not. If none of the elements in the
section should be shown, the whole section can be hidden.

Bug: 449919991
Test: atest Enable2gPreferenceControllerTest
Test: atest CellularSecurityPreferenceControllerTest
Flag: EXEMPT BUGFIX

Change-Id: I365b3bc96bb278db2a3b5d3ccf9ac66bbd041e5c