Fix dialog not showing when update is done on external display
When resolution change is done on external display, the display
gets reconfigured and the launched dialog is immediately dismissed,
issuing cancellation on the resolution change itself. So there
needs to be some "state manager" informing that there is a pending
change, which persists through fragment lifecycle. There are multiple
options to fix this:
1. Use viewmodel and observe the livedata
This works, but in this fragment there are multiple sources of updates:
- DisplayListener
- ViewModel
Both will be triggered at the same time causing flickers and possibly
race condition
2. Use viewmodel only to persist state
Rely on DisplayListener as the main update pipeline. In update(), check
the latest viewmodel data to see if there's any pending confirmation to
show/reshow the dialog

Bug: 454516501
Test: atest ResolutionPreferenceFragmentTest
Test: atest ResolutionChangeDialogFragmentTest
Flag: com.android.settings.flags.enable_resolution_confirm_dialog_bugfix
Change-Id: I77da68f2af4e1d8948b4b6fb04dea8f742797b1c