[L10N] Create separate strings for unavailable state
It's unclear when a placeholder string should be lower cases when appear
in a sentence. Though go/android-naming mentioned strings should be
lowercase when used in the middle of a sentence because they are generic
descriptive names and not branded names, the same rule doesn't apply to
other locales like de, fr, ru.

Bug: 436455949
Test: atest SettingsRoboTests
Test: manual
Flag: EXEMPT BUGFIX
Change-Id: If4fbbbc17a921e132b70677e00e26688faeb5a82