Add satellite apps section to the satellite landing page.
This change introduces a new `ComposePreference` in the satellite
landing page layout to display a list of applications compatible with
satellite. The list of apps is dynamically populated based on whether
the device is in an LTE-based or NB-NTN satellite scenario.

The `SatelliteAppsRepository` is introduced to manage and provide these
lists of apps. It reads app configurations from `DeviceConfig`,
including default values for both NB-NTN and LTE scenarios. Hardcoded
entries for "Emergency SOS" and "Settings" are also included.

The `SatelliteLandingPageFragment` reads the app list from
`SatelliteLandingPageViewModel`, which loads the satellite app info from
`SatelliteAppsRepository`.

Subsequent changes will implement the getSatelliteDataOptimizedApps API
to fill the apps list for the LTE Landing Page.

Test: atest packages/apps/Settings/tests/robotests/src/com/android/settings/network/telephony/satellite/quicksettings/SatelliteAppsRepositoryTest.kt, atest packages/apps/Settings/tests/robotests/src/com/android/settings/network/telephony/satellite/quicksettings/SatelliteLandingPageFragmentTest.kt, atest packages/apps/Settings/tests/robotests/src/com/android/settings/network/telephony/satellite/quicksettings/SatelliteAppItemTest.kt, atest packages/apps/Settings/tests/robotests/src/com/android/settings/network/telephony/satellite/quicksettings/SatelliteLandingPageViewModelTest.kt
Bug: 434793872
Flag: com.android.settings.flags.enable_satellite_tile
Change-Id: Ib51a6345cd91d35f3c26b00dbde2f46b9bba7163