Add a warning asking user to backup device data
The "Experimental: Enable ANGLE" switch is aimed for dogfood users
and developers to test ANGLE as the system driver. On a released
commercial android device, user should not play around with the switch,
as this can brick the device into unusable state and need a factory
reset to fix. Add a warning to ask user to backup their device data
so that they are aware of the potential risk, and if they do decide
to toggle the switch, they can mitigate the risk by backing up their
device data.

This change should be backported to earlier Android release.

Flag: EXEMPT BUGFIX
Bug: b/270994705
Test: make, flash to Pixel 9
Change-Id: I10501b6fdee67621151aa2b206586f2024a7f95e