[ScreenOnTime] Add pre-process while combining the app events with device events.
For the top activity with screen-off events, the UsageStatsManager usually record screen-off device event first, then record the stop events for app activities.

We reverse the order of the adjacent screen-off event and stop events if they happened in a short duration. This is prepared for the usage period calculation algorithm later.

- Before:| App start | Screen off | App stop |
In this case, both app start/stop event will be considered as unmatched events and closed as the incomplete duration(+30s).

- After: | App start | App stop | Screen off |
After processing, the start+stop can get the best match and generate a usage period.

Bug: 445131736
Test: atest DataProcessorTest
Flag: EXEMPT BUGFIX
Change-Id: I670afd093154c003fa768cb815705c7e16dfcefd