Add config to hide pattern lock option
This change introduces a new config to allow hiding the pattern
lock screen option from Settings and the Setup Wizard.

When this flag is enabled:
- The "Pattern" option will not be visible in the lock screen type
  selection list.
- The checkbox on the pattern confirmation screen, which allows
  importing a pattern from another device during Quick Start will
  be hidden.

This provides a mechanism for device configurations, such as
Android Desktop, to disable the use of new pattern locks while not
affecting users who already have a pattern set.

Bug: 440006118
Test: atest ChooseLockGenericControllerTest
Test: atest ConfirmLockPatternTest
Test: manual test on device
Flag: NONE new config will only be enabled for desktop
Change-Id: I09a9f6366c9987f0a2a2f12d0b3c320623e68088