Refactor SatelliteTileStateReceiver with DeviceConfig check
Add a DeviceConfig flag to enable or disable the satellite tile on a
per-device basis. This allows for more granular control over the
feature's rollout.

The new check is performed after the hardware feature check to ensure
that the receiver fails as early as possible on unsupported devices.

Refactor the intent action check in onReceive to use a when statement
for improved readability.

These checks should also help with performance metrics regressions on
satellite unsupported devices.

Bug: 434793872, 452169645, 460900000
Flag: com.android.settings.flags.enable_satellite_tile
Test: atest SatelliteTileStateReceiverTest
Change-Id: I71132acb4325a2e9bdacca777ec4fd9797f15742